Add files via upload
It is a voluntary and non-governmental organization. It maintains Online library of blood 
donors in India. Sometimes Doctors and Blood bank project have to face the difficulty in finding 
the blood group Donors at right time. Help Line has attempted to provide the answer by taking upon 
itself the task of collecting Blood bank project nationwide for the cause and care of people in need.
At any point of time the people who are in need can reach the donors through our search facility. 
By mobilizing people and organization who desire to make a difference in the lives of people in 
need. On the basis of humanity, everyone is welcome to register as a blood donor.
The Blood Alliance (TBA) is an app-based system that is designed to store, process, retrieve and 
analyse information concerned with the administrative and inventory management within a blood 
bank. This project aims at maintaining all the information pertaining to blood donors, different blood 
groups available in each blood bank and help them manage in a better way. Aim is to provide 
transparency in this field, make the process of obtaining blood from a blood bank hassle free and 
corruption free and make the system of blood bank management effective.